fix handling of whitelist
The packages  specified in  the whitelist  are now  treated correctly.
Thus  a package  is  only considered  to be  exported  if its  initial
characters match one of the  packages within the whitelist. Previously
this was not the case and thus package names as in example "db" caused
the such package "gdbm" also to be exported.

Furthermore  the  extraction script  was  extended  to further  reduce
the  packages in  whitelist even  more.  This is  necessary since  the
differentation between meta packages not installed on the platform and
real target packages is not really  reliable and we are not allowed to
deliver packages with GPLv3 in.  These are filtered out explicitly now
before the clojure application is started.

Signed-off-by: Otto Linnemann <otto.linnemann@peiker.de>