exclude gettext package from whitelist
The package  gettext is mainly  used as  meta package and  under GPL3.
Since it is not annotated correctly, pub-oss has no information wether
it is  delivered with the platform  image or not. We  therefore remove
this package from the whitelist and  ensure in this way that this GPL3
package is not distributed. Please  do carefully check whether gettext
is part of the platform image nevertheless.

Signed-off-by: Otto Linnemann <otto.linnemann@peiker.de>